certimate-backend/requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
python-multipart==0.0.20
pillow==10.4.0
pandas==2.2.2
pytesseract==0.3.10
pdf2image==1.16.3
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic-settings==2.6.1
requests==2.32.3
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.158.0
email-validator==2.2.0
dnspython
psutil==6.1.0
